+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPageDefault/EnrollmentPageDefault.component.js
@@ -45,6 +45,7 @@ const getStyles = ({ typography }) => ({
 export const EnrollmentPageDefaultPlain = ({
     program,
     teiId,
+    orgUnitId,
     events,
     enrollmentId,
     stages,
@@ -88,7 +89,7 @@ export const EnrollmentPageDefaultPlain = ({
                         emptyText={i18n.t('No feedback for this enrollment yet')}
                     />
                 )}
-                <WidgetProfile teiId={teiId} programId={program.id} />
+                <WidgetProfile teiId={teiId} programId={program.id} showEdit orgUnitId={orgUnitId} />
                 {enrollmentId !== 'AUTO' && <WidgetEnrollment
                     teiId={teiId}
                     enrollmentId={enrollmentId}