+++ src/core_modules/capture-core/components/WidgetProfile/DataEntry/ProgramRules/index.js
@@ -1,0 +1,3 @@
+// @flow
+export { buildRulesContainer } from './rulesContainer';
+export { getRulesActionsForTEI } from './getRulesActionsForTEI';