+++ src/core_modules/capture-core/components/WidgetProfile/DataEntry/dataEntry.types.js
@@ -1,0 +1,19 @@
+// @flow
+
+export type PlainProps = {|
+    dataEntryId: string,
+    itemId: string,
+    trackedEntityName: string,
+    saveAttempted: boolean,
+    formFoundation: any,
+    onCancel: () => void,
+    onSave: () => void,
+    onUpdateFormField: () => void,
+|};
+
+export type Props = {|
+    programAPI: any,
+    orgUnitId: string,
+    onCancel: () => void,
+    clientAttributesWithSubvalues: Array<any>,
+|};