+++ src/core_modules/capture-core/components/WidgetProfile/hooks/index.js
@@ -1,0 +1,3 @@
+export { useProgram } from './useProgram';
+export { useTrackedEntityInstances } from './useTrackedEntityInstances';
+export { useClientAttributesWithSubvalues } from './useClientAttributesWithSubvalues';